change print_num

+++ ft_printf/ft_printf/ft_printf.h
@@ -23,7 +23,7 @@ void	print_type_xX(info *test, unsigned int num, size_t *count);
 void	print_type_u(info *test, unsigned int num, size_t *count);
 void	print_type_id(info *test, int num, size_t *count);
 void	print_hex(char type, size_t num);
-void	print_num(long num);
+void	print_num(unsigned int num);
 void	free_box(info *test);
 
 #endif
